placed handlers in correct place
