+++ auth/google.go
@@ -6,7 +6,6 @@ import (
 	"encoding/base64"
 	"fmt"
 	"io"
-	"log"
 	"net/http"
 	"os"
 	"time"
@@ -15,7 +14,7 @@ import (
 	"golang.org/x/oauth2/google"
 )
 
-var googleOauthConfig = &oauth2.Config{
+var GoogleOauthConfig = &oauth2.Config{
 	RedirectURL: "http://localhost:8080/auth/google/callback",
 	ClientID: os.Getenv("GOOGLE_OAUTH_CLIENT_ID"),
 	ClientSecret: os.Getenv("GOOGLE_OAUTH_CLIENT_SECRET"),
@@ -25,29 +24,7 @@ var googleOauthConfig = &oauth2.Config{
 
 const oauthGoogleUrlAPI = "https://www.googleapis.com/oauth2/v2/userinfo?access_token="
 
-func OauthGoogleLogin(w http.ResponseWriter, r *http.Request){
-	oauthState := generateStateOauthCookie(w)
-	u := googleOauthConfig.AuthCodeURL(oauthState)
-	http.Redirect(w, r, u, http.StatusTemporaryRedirect)
-}
-
-func OauthGoogleCallback(w http.ResponseWriter, r *http.Request){
-	oauthState, _ := r.Cookie("oauthstate")
-	if r.FormValue("state") != oauthState.Value {
-		log.Println("invalid oauth state - this has been tampered with")
-		http.Redirect(w, r, "/", http.StatusTemporaryRedirect)
-		return
-	}
-	data, err := getUserComingFromGoogle(r.FormValue("code"))
-	if err != nil {
-		log.Println(err.Error())
-		http.Redirect(w, r, "/", http.StatusTemporaryRedirect)
-		return
-	}
-
-	fmt.Fprintf(w, "%s", data)
-}
-func generateStateOauthCookie(w http.ResponseWriter) string {
+func GenerateStateOauthCookie(w http.ResponseWriter) string {
 	var expiration = time.Now().Add(20 * time.Minute)
 	b := make([]byte, 16)
 	rand.Read(b)
@@ -56,8 +33,8 @@ func generateStateOauthCookie(w http.ResponseWriter) string {
 	http.SetCookie(w, &cookie)
 	return state
 }
-func getUserComingFromGoogle(code string) ([]byte, error) {
-	token, err := googleOauthConfig.Exchange(context.Background(), code)
+func GetUserComingFromGoogle(code string) ([]byte, error) {
+	token, err := GoogleOauthConfig.Exchange(context.Background(), code)
 	if err != nil {
 		return nil, fmt.Errorf("code exchange gone wrong!! %s", err.Error())
 	}
