+++ main.go
@@ -3,7 +3,6 @@ package main
 import (
 	"net/http"
 
-	"github.com/anyuan-chen/urlshortener/server/auth"
 	"github.com/anyuan-chen/urlshortener/server/handler"
 	"github.com/anyuan-chen/urlshortener/server/store"
 	"github.com/gorilla/mux"
@@ -14,8 +13,8 @@ func main() {
 	store.InitializeStore()
 	r.HandleFunc("/create/{url}", handler.CreateShortUrl).Methods("POST")
 	r.HandleFunc("/redirect/{url}", handler.Redirect).Methods("GET")
-	r.HandleFunc("/auth/google/login", auth.OauthGoogleLogin)
-	r.HandleFunc("/auth/google/callback", auth.OauthGoogleCallback)
+	r.HandleFunc("/auth/google/login", handler.OauthGoogleLogin)
+	r.HandleFunc("/auth/google/callback", handler.OauthGoogleCallback)
 	http.ListenAndServe(":8080", r)
 
 }
